changed numpy back to old way, removed unecessary serialization step

+++ examples/template_module.py
@@ -17,7 +17,7 @@ class MyNestedParameters(mm.Schema):
                          'description': 'name of vector'})
     increment = mm.fields.Int(required=True, metadata={
                               'description': 'value to increment'})
-    array = NumpyArray(dtype=np.int, required=True, metadata={
+    array = NumpyArray(dtype=np.uint8, required=True, metadata={
                        'description': 'array to increment'})
     write_output = Boolean(required=False, default=True)
